requirements.txt
amqp==5.3.1
asgiref==3.10.0
async-timeout==5.0.1
billiard==4.2.2
celery==5.5.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
django==4.2.25
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.1
kombu==5.5.4
packaging==25.0
prompt-toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
